"Fix" the red background issue in asciicasts
This was here for debugging purposes. It's now blue. So, it's still there,
but it blends in there. The reason it needs to be there is so that when
the page is loading, before the embed has loaded, it provides a placeholder.
If it's not there, it results in a janky white flash on load.