Fix type-checking errors when importing react-dropzone-uploader
Include the TS file from `react-dropzone-uploader` in a TS project that
has altered `compilerOptions` in an attempt to ignore the TS errors.

This is because tsc type-checks the TS files in
`node_modules/react-dropzone-uploader` and reports errors in them. They
are TS files (not `d.ts`), so `skipLibCheck` does not affect them.
`exclude` is not something to use here, because the files are imported
directly, so `exclude` does not affect whether they are type-checked or
not.