Fix `withMUITheme` typing
`withMUITheme` accepted `any` and returned a component that was not
fully typed (the main problem was that it did not specify `props`).

This commit changes the HOC typing to forward the type of the passed
component as the return type.

This makes it so that types for props are correctly exposed on the
returned component.

It is important when using `withMUITheme` in tests to wrap a component
and then pass props to it.

Related to CU-1m7bk32