Remove unused and probably wrong example of CSS prop
Remove the demo of the `css` prop. The code does not seem to be imported
at all. Looking at the TypeScript errors after removing `@ts-nocheck`,
the code likely did not work, because the color is accessed immediately
on the theme (`theme.primary.main`) instead of accessing the `palette`
first (`theme.palette.primary.main`).

The code is dead weight at this point and there is little value in
keeping it.

It had an `eslint-disable`, which is why it was found when modifying
ESLint setup.

- CU-20jjp6y