Add component names and simplify mdx components
Specifying the component names is a hint for React Hooks ESLint plugin
to treat those functions as components and allow using hooks inside of
them.

Moreover, this commit refactors the way props are spread to pass all the
props to the underlying component in a single batch (`props`), instead
of extracting `children` separately.