Rename session ID cookie as well
Note: This code is in the connect page, which is not currently
being rebuilt and redeployed. That means the change will not actually
affect live code, and so the code that checks if the session ID exists
will only be checking the old 'sgr.session.id' cookie instead of
'sgr.session.id.v2'.

For this reason, until we deploy the new /connect page, we are also
changing the catalog code to set two session cookies: one for v2
and one for the original.