Stretch colored icons
Fix icon sizes when the `color` prop is specified to stretch according
to the specified `size`.

The `backgroundSize` style has an effect only when the `color` prop is
not specified, because then, the whole SVG is used as a background and
then resized according to `backgroundSize`.

When the `color` prop is specified, the icon SVG is used as a mask,
which is not resized by default. This means the icon had the same
dimensions as defined in the SVG data URI.

The mask-size CSS property fixes this problem by having the mask cover
the whole element (according to its `size`). This way the icon has the
same size regardless of whether it has a `color` specified or not.

See https://developer.mozilla.org/en-US/docs/Web/CSS/mask-size

https://app.clickup.com/t/20xxby3